Zero's and NA's removed from lifetable
The code inside the validity method for lifetable to remove zero's and
NA's does not have the intended effect.

Created a constructor method to remove 0's and NA's and removed the code
from the validity method.

Added test file for object initialization to check the removal of NA's
and 0's.