Fix fractional pxt
Function pxt returned zero when x + t was greater than
or equal to the last positive entry in lx.  As a consequence,
under linear interpolation, the calculation of Axn with k > 1
did not return correct values.

For linear interpolation, when x + t is between the last possible
age (with lx >0) and the next age (lx = 0), pxt should return
positive results.  In this case we are interpolating lx between
the last lx > 0 and the next one lx = 0.