Fix calculation of number of periods, n, in Axn
Function Axn was not returning correct values because it
was ignoring the last couple of entries in its calculation.

The current definition of omega is that it is equal to the
age corresponding to the last entry in lx > 0.  In a table
where ages <- 0:9 we have that omega = 9.  For a person
age x = 7, then there would be three intervals with a positive
probability of death; (7,8], (8,9], and (9,10].  Therefore,
the calculation of n should be n = omega + 1 - x = 9 + 1 - 7 = 3.