Fix soa08 and soa08Act tables, add tests, update documentation
The soa08 and soa08Act tables did not match the published
table in the book Actuarial Mathematics.  Created new data
objects with the correct tables.  As explained it the
exercises to chapter 3 of the book this Illustrative Life
Table is a blend of ad hoc entries for ages 0 to 12 and
then Makeham's mortality law for ages 13 all the way to 140.
The published table only shows up to age 110.  By looking at
the annuity values for age 110 you can see that the table
extends beyond that.

Update the documentation entries for both data objects.

Added tests for the functions axn, Axn, and Axn with power = 2 to
match the values published in the book.

Also added tests for multiple life functions axyzn, Axyzn, and
Axyzn with power = 2 to match the values published in the book.